Fix try to delete file while it does not exist anymore (see #2102 and #1713) in reference to this issue

+++ codec-http/src/main/java/io/netty/handler/codec/http/multipart/AbstractDiskHttpData.java
@@ -226,9 +226,10 @@ public abstract class AbstractDiskHttpData extends AbstractHttpData {
     @Override
     public void delete() {
         if (! isRenamed) {
-            if (file != null) {
+            if (file != null && file.exists()) {
                 file.delete();
             }
+            file = null;
         }
     }
 
@@ -307,6 +308,9 @@ public abstract class AbstractDiskHttpData extends AbstractHttpData {
     public boolean renameTo(File dest) throws IOException {
         if (dest == null) {
             throw new NullPointerException("dest");
+        }
+        if (file == null) {
+            throw new IOException("No file defined so cannot be renamed");
         }
         if (!file.renameTo(dest)) {
             // must copy
